+++ src/main/java/umc/stockoneqback/field/domain/company/CompanyRepository.java
@@ -1,0 +1,9 @@
+package umc.stockoneqback.field.domain.company;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface CompanyRepository extends JpaRepository<Company, Long> {
+    Optional<Company> findByName(String name);
+}
