+++ src/main/java/umc/stockoneqback/field/domain/store/PartTimerRepository.java
@@ -1,0 +1,11 @@
+package umc.stockoneqback.field.domain.store;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import umc.stockoneqback.user.domain.User;
+
+import java.util.Optional;
+
+public interface PartTimerRepository extends JpaRepository<PartTimer, Long> {
+    Optional<PartTimer> findByPartTimer(User user);
+    void deleteByPartTimer(User user);
+}
