--- src/main/java/umc/stockoneqback/role/domain/company/CompanyRepository.java
@@ -1,9 +1,0 @@
-package umc.stockoneqback.role.domain.company;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface CompanyRepository extends JpaRepository<Company, Long> {
-    Optional<Company> findByName(String name);
-}
