+++ src/test/java/umc/stockoneqback/fixture/StoreFixture.java
@@ -2,7 +2,7 @@ package umc.stockoneqback.fixture;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import umc.stockoneqback.role.domain.store.Store;
+import umc.stockoneqback.field.domain.store.Store;
 import umc.stockoneqback.user.domain.User;
 
 @Getter
